// src/pages/notes/_lib/api-types.ts
export interface components {
    schemas: {
// ...
        VersionSummaryOut: {
            /** Authorid */
            authorId: string;
            /** Createdat */
            createdAt: number;
            /** Sizebytes */
            sizeBytes: number;
            /** Version */
            version: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}